AI HackDay (by Swaacha Tech company)/EpicEcho/requirements.txt
google.generativeai
streamlit
requests